backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.9
openai==1.12.0
python-dotenv==1.0.1
chromadb==0.4.22
pydantic==2.6.1
python-multipart==0.0.9